requirements.txt
discord.py==2.7.0
python-dotenv==1.2.2
spotipy==2.25.2
SQLAlchemy==2.0.48
